# Improve performance of the scheduled task